Add intentional idling, fix bounds
One of the randomly generated numbers now signifies idling
(no movement)

When the pet hit a bound, it wouldn't move for that turn at all.
Now, it will always move in the opposite direction from the wall.

The bounds were flipped for the width and height. This is now
fixed